changing 'a' to have its own type in oputils. still not work properly.

+++ tests/expm_multiply_parallel_test.py
@@ -73,8 +73,8 @@ def test_ramdom_int_matrix(N=3500,ntest=10,seed=0):
 	i = 0
 	while(i<ntest):
 		print("testing random integer matrix {}".format(i+1))
-		data_rvs = lambda n:np.random.randint(-100,100,size=n)
-		A = random(N,N,density=np.log(N)/N,data_rvs=data_rvs).astype(np.int8)
+		data_rvs = lambda n:np.random.randint(-100,100,size=n).astype(np.int8)
+		A = random(N,N,density=np.log(N)/N,data_rvs=data_rvs,dtype=np.int8)
 		A = A.tocsr()
 
 		v = np.random.normal(0,1,size=N) + 1j * np.random.normal(0,1,size=N)
